refactor: PainteraAlerts location, and extract some related logic to separate files
