+++ src/main/java/org/janelia/saalfeldlab/paintera/ui/dialogs/open/menu/thresholded/ThresholdedRawSourceStateOpenerDialog.java
@@ -30,7 +30,7 @@ import org.janelia.saalfeldlab.paintera.state.SourceInfo;
 import org.janelia.saalfeldlab.paintera.state.SourceState;
 import org.janelia.saalfeldlab.paintera.state.ThresholdingSourceState;
 import org.janelia.saalfeldlab.paintera.state.raw.ConnectomicsRawState;
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts;
+import org.janelia.saalfeldlab.paintera.ui.dialogs.PainteraAlerts;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
