+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/config/BookmarkConfigNode.kt
@@ -20,7 +20,7 @@ import org.janelia.saalfeldlab.fx.ui.MarkdownPane
 import org.janelia.saalfeldlab.fx.ui.NumericSliderWithField
 import org.janelia.saalfeldlab.fx.util.InvokeOnJavaFXApplicationThread
 import org.janelia.saalfeldlab.paintera.Paintera
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
+import org.janelia.saalfeldlab.paintera.ui.dialogs.PainteraAlerts
 import org.slf4j.LoggerFactory
 import java.lang.invoke.MethodHandles
 import java.util.Optional
