+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/config/MultiBoxOverlayConfigNode.kt
@@ -8,11 +8,10 @@ import javafx.geometry.Pos
 import javafx.scene.Node
 import javafx.scene.control.*
 import javafx.scene.layout.HBox
-import javafx.stage.Modality
 import javafx.util.Callback
 import org.janelia.saalfeldlab.fx.extensions.TitledPaneExtensions
 import org.janelia.saalfeldlab.fx.ui.NamedNode
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
+import org.janelia.saalfeldlab.paintera.ui.dialogs.PainteraAlerts
 
 class MultiBoxOverlayConfigNode() {
