+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/config/PainteraDirectoriesConfig.kt
@@ -19,7 +19,7 @@ import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.get
 import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.set
 import org.janelia.saalfeldlab.paintera.serialization.PainteraSerialization
 import org.janelia.saalfeldlab.paintera.ui.FontAwesome
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
+import org.janelia.saalfeldlab.paintera.ui.dialogs.PainteraAlerts
 import org.scijava.plugin.Plugin
 import java.lang.reflect.Type
