+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/config/input/KeyAndMouseConfigNode.kt
@@ -27,7 +27,7 @@ import org.janelia.saalfeldlab.paintera.control.modes.ControlMode
 import org.janelia.saalfeldlab.paintera.control.modes.NavigationTool
 import org.janelia.saalfeldlab.paintera.state.SourceInfo
 import org.janelia.saalfeldlab.paintera.state.SourceState
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
+import org.janelia.saalfeldlab.paintera.ui.dialogs.PainteraAlerts
 import java.util.Locale
 
 class KeyAndMouseConfigNode(
