+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/state/SourceStateCompositePane.kt
@@ -16,7 +16,7 @@ import org.janelia.saalfeldlab.paintera.composition.ARGBCompositeAlphaYCbCr
 import org.janelia.saalfeldlab.paintera.composition.AlphaCopy
 import org.janelia.saalfeldlab.paintera.composition.Composite
 import org.janelia.saalfeldlab.paintera.composition.CompositeCopy
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
+import org.janelia.saalfeldlab.paintera.ui.dialogs.PainteraAlerts
 
 typealias FXC = FXCollections
 typealias ARGBComposite = Composite<ARGBType, ARGBType>
