+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/ui/dialogs/ExportSourceDialog.kt
@@ -28,7 +28,6 @@ import org.janelia.saalfeldlab.paintera.paintera
 import org.janelia.saalfeldlab.paintera.serialization.GsonExtensions.get
 import org.janelia.saalfeldlab.paintera.state.SourceStateBackendN5
 import org.janelia.saalfeldlab.paintera.state.label.ConnectomicsLabelState
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
 import org.janelia.saalfeldlab.paintera.ui.menus.PainteraMenuItems
 import org.janelia.saalfeldlab.util.n5.N5Helpers.MAX_ID_KEY
 import kotlin.jvm.optionals.getOrNull
