+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/ui/dialogs/KeyBindingsDialog.kt
@@ -6,7 +6,6 @@ import javafx.stage.Modality
 import org.janelia.saalfeldlab.paintera.config.input.KeyAndMouseConfigNode
 import org.janelia.saalfeldlab.paintera.paintera
 import org.janelia.saalfeldlab.paintera.properties
-import org.janelia.saalfeldlab.paintera.ui.PainteraAlerts
 import org.slf4j.LoggerFactory
 
 internal object KeyBindingsDialog {
